% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_d_from_AUC.R
\name{calc_d_from_AUC}
\alias{calc_d_from_AUC}
\title{Calculate Cohen's d from AUC}
\usage{
calc_d_from_AUC(AUC)
}
\arguments{
\item{AUC}{AUC-value(s)}
}
\value{
Cohen's d value(s)
}
\description{
This function transforms AUC values into Cohen's d assuming
equal group sizes. The formula used is from Table 1 in Ruscio (2008).
}
\examples{
calc_d_from_AUC(c(0.5, 0.7, 0.9))
}
