% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_d_from_R2.R
\name{calc_d_from_R2}
\alias{calc_d_from_R2}
\title{Calculate Cohen's d from R-Squared}
\usage{
calc_d_from_R2(R2)
}
\arguments{
\item{R2}{R-squared value(s)}
}
\value{
Cohen's d value(s)
}
\description{
This function transforms R-squared values into Cohen's d assuming
equal group sizes. The formula used is from Table 1 in Ruscio (2008).
R-square is treated as the square of r.
}
\examples{
calc_d_from_R2(c(0, 0.1, 0.2))
}
