% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_logit_from_p.R
\name{calc_logit_from_p}
\alias{calc_logit_from_p}
\title{Caclulate logit (log odds) from probability}
\usage{
calc_logit_from_p(probability, offset_for_01 = NULL)
}
\arguments{
\item{probability}{A numerical vector (or single number) with probabilities
between 0 and 1.}

\item{offset_for_01}{A decimal number to add to probability 0 or subtract from 1 to make
calculation of logit possible.}
}
\value{
A numerical vector (or single number) representing the logit (log
  odds).
}
\description{
Caclulate logit (log odds) from probability
}
\examples{
calc_logit_from_p(0.2)

probabilities <- c(0.00, 0.07, 0.12,0.74, 1.00)
calc_logit_from_p(probabilities, offset_for_01 = 0.001)

}
