% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_p_from_logit.R
\name{calc_p_from_logit}
\alias{calc_p_from_logit}
\title{Calculate probability from  logit (log odds)}
\usage{
calc_p_from_logit(logit)
}
\arguments{
\item{logit}{A numerical vector (or single number) with logits.}
}
\value{
A numerical vector (or single number) representing the probability
  odds).
}
\description{
Calculate probability from  logit (log odds)
}
\examples{
log_odds <- c(log(0.01/0.99), log(0.25/0.75), log(1/1), log(99/1))
calc_p_from_logit(logit = log_odds)

}
