% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{descriptive_stats_cat}
\alias{descriptive_stats_cat}
\title{Descriptive statistics for categorical variables}
\format{An object of class \code{data.frame} with 233 rows and 8 columns.}
\source{
data-raw/descriptives.Rmd
}
\usage{
descriptive_stats_cat
}
\description{
#' @format A data frame with 233 rows and 8 variables: \describe{
\item{variable}{Variable name as used in R scripts}
\item{var_level}{Level on the variable to which the frequencies and percentages refer}
\item{no_reoffence_freq}{Frequency in the group of individuals that did not
reoffend}
\item{reoffence_nonviolent_freq}{Frequency in the group of individuals
that reoffended with a non-violent crime}
\item{reoffence_violent_freq}{Frequency in the group of individuals
that reoffended with assault or homicide}
\item{no_reoffence_perc}{Percentage in the group of individuals that
did not reoffend}
\item{reoffence_nonviolent_perc}{Percentage in the group of individuals
that reoffended with a non-violent crime}
\item{reoffence_violent_perc}{Percentage in the group of individuals
that reoffended with assault or homicide} }
}
\keyword{datasets}
