% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{descriptive_stats_num}
\alias{descriptive_stats_num}
\title{Descriptive statistics for numerical variables}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20 rows and 7 columns.}
\source{
data-raw/descriptives.Rmd
}
\usage{
descriptive_stats_num
}
\description{
#' @format A data frame with 20 rows and 7 variables: \describe{
\item{variable}{Variable name as used in R scripts}
\item{no_reoffence_mean}{Mean for the group of individuals that did not
reoffend}
\item{reoffence_nonviolent_mean}{Mean for the group of individuals
that reoffended with a non-violent crime}
\item{reoffence_violent_mean}{Mean for the group of individuals
that reoffended with assault or homicide}
\item{no_reoffence_sd}{Standard deviation for the group of individuals that
did not reoffend}
\item{reoffence_nonviolent_sd}{SD for the group of individuals
that reoffended with a non-violent crime}
\item{reoffence_violent_sd}{SD for the group of individuals
that reoffended with assault or homicide} }
}
\keyword{datasets}
