% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{model_AUCs_training_set}
\alias{model_AUCs_training_set}
\title{Discrimination of models in training set.y}
\format{A list with 6 elements of class 'resamples' created by
'caret::resamples'}
\source{
Scripts for how the results were created are stored in /data-raw
}
\usage{
model_AUCs_training_set
}
\description{
.
The 'values' element contains a data frame with 100 rows of the results in
the 10 by 10 fold cross-validation. For each model the ROC(AUC),
is recorded along with sensitivity and specificity for decision cutoff of 0.5.
That gives us 24 model x 3 measures = 72 columns. With the resample numbers
that is 73 columns.
}
\seealso{
caret::resamples
}
\keyword{datasets}
