% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_perfs_training_set1000}
\alias{model_perfs_training_set1000}
\title{Model performances in the training set based on 1000 resamples}
\format{An object of class \code{resamples} of length 6.}
\source{
The data object is created in
  "data-raw/06_model_perfs_training_set.Rmd". See also
  "data-raw/00_About_data_raw.Rmd"
}
\usage{
model_perfs_training_set1000
}
\description{
This a data object of class \code{resamples} generated by
\code{caret::resamples} that contains the model performance of all models in
the training set based on 250 resamples of a 4-fold cross-validation. Model names can be looked up
\code{model_grid}.
}
\details{
The performance metrics are:
- d_AUC: Cohen's d based on a naive transformation from Area Under Curve
- E_O_ratio
- logLoss
- McF_R2: McFadden's Pseudo R^2
- ROC: Area Under Curve from a receiver operating characteristic curve.
}
\keyword{datasets}
