% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{predictions}
\alias{predictions}
\title{Predictions on test set}
\format{A data frame with 374 observation of 24 variables
\describe{
  \item{}{}
  \item{}{}
  ...
}}
\source{
Database
}
\usage{
predictions
}
\description{
Estimated probabilities of recidivism for the 374 individuals in the test set.
Variables starting with 'G_' refer to probabilities of general recidivism and
Variables starting with 'V_' refer to probabilities of violent recidivism
}
\keyword{datasets}
