% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{static_preds}
\alias{static_preds}
\title{Static predictors}
\format{A vector of 24 variable names:
\describe{
  \item{ageFirstSentence_mr}{Age at first sentence}
  \item{ageFirst_missing}{Was the information about 'ageFirstSentence' missing?}
  ...
}}
\source{
Database
}
\usage{
static_preds
}
\description{
The static predictors used in the study
}
\keyword{datasets}
