% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variable_table}
\alias{variable_table}
\title{Variable table}
\format{A data frame with 114 rows and 5 variables:
\describe{
  \item{Variable}{Variable name as used in R scripts}
  \item{Label}{Short descriptive label}
  \item{Role}{The role the variable plays in the analyses.
  Can be outcome, predictor_static, predictor_dynamic, predictor_term, or descriptive}
  \item{Desriptives}{The type of descriptive statistics to produce for the variable}
  \item{Comment}{Additional comments}

}}
\source{
variable_table.csv compiled by Benny Salo
}
\usage{
variable_table
}
\description{
Variable names, labels, the role they play in the analyses and comments.
For RITA-items comments this includes the full wording of the item translated
from english by the study authors.
}
\keyword{datasets}
