% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_p_from_logit.R
\name{calc_p_from_logit}
\alias{calc_p_from_logit}
\title{Calculate probability from  logit (log odds)}
\usage{
calc_p_from_logit(logit)
}
\arguments{
\item{logit}{A numerical vector (or single number) with logits.}
}
\value{
A numerical vector (or single number) representing the probability
  odds).
}
\description{
Calculate probability from  logit (log odds)
}
\examples{
calc_logit_from_p(0.2)

probabilities <- c(0.00, 0.07, 0.12,0.74, 1.00)
calc_logit_from_p(probabilities, offset_for_01 = 0.001)

}
